gpt-sovits-mlx/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-sovits-mlx"
version = "0.1.0"
description = "GPT-SoVITS TTS on Apple Silicon using MLX and CoreML"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "OminiX-MLX Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mlx>=0.10.0",
    "mlx-lm>=0.10.0",
    "numpy>=1.24.0",
    "safetensors>=0.4.0",
    "tokenizers>=0.15.0",
    "huggingface-hub>=0.19.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
coreml = [
    "coremltools>=7.0",
    "ane-transformers>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "gpt-sovits-mlx[coreml,dev]",
]

[project.scripts]
gpt-sovits-mlx = "gpt_sovits_mlx.cli:main"

[project.urls]
Homepage = "https://github.com/oxideai/mlx-rs"
Repository = "https://github.com/oxideai/mlx-rs"

[tool.setuptools.packages.find]
where = ["."]
include = ["python*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
